+++ bare-metal/fedora-coreos/kubernetes/outputs.tf
@@ -2,3 +2,9 @@ output "kubeconfig-admin" {
   value = module.bootstrap.kubeconfig-admin
 }
 
+# Outputs for debug
+
+output "assets_dist" {
+  value = module.bootstrap.assets_dist
+}
+