+++ digital-ocean/container-linux/kubernetes/outputs.tf
@@ -54,3 +54,9 @@ output "vpc_id" {
   value       = digitalocean_vpc.network.id
 }
 
+# Outputs for debug
+
+output "assets_dist" {
+  value = module.bootstrap.assets_dist
+}
+